% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_date.R
\name{IsDate}
\alias{IsDate}
\title{IsDate}
\usage{
IsDate(mydate, date.format = "\%Y-\%m-\%d")
}
\arguments{
\item{mydate}{list of dates}
}
\value{
true or false
}
\description{
A function to detect the desired date format
\cite{\link{median_data_computation function}}.
}
