% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_median_data_computation.R
\name{median_data_computation}
\alias{median_data_computation}
\title{median_data_computation}
\usage{
median_data_computation(
  input,
  meta,
  new_format = TRUE,
  row_wise = TRUE,
  output
)
}
\arguments{
\item{input}{input file (.csv)}

\item{meta}{metafile}

\item{new_format}{Logical value specifying if new format. Default = TRUE.}

\item{row_wise}{Logical value specifying if the output required is in wide format}

\item{output}{output file location}
}
\value{
list containing input data, meta data, median calculated from the traits.
}
\description{
Median calculation
}
