% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_median_data_computation.R
\name{median_data_computation}
\alias{median_data_computation}
\title{median_data_computation}
\usage{
median_data_computation(
  data = NULL,
  meta_data = NULL,
  input_file = NULL,
  meta_file = NULL,
  new_format = TRUE,
  row_wise = TRUE,
  output
)
}
\arguments{
\item{data}{plant eye data with specific column and column names required.}

\item{meta_data}{meta data.}

\item{input_file}{path to the input plant eye data (.csv)}

\item{meta_file}{path to the meta data (.csv)}

\item{new_format}{Logical value specifying if new format. Default = TRUE.}

\item{row_wise}{Logical value specifying if the output required is in wide format}

\item{output}{output file location}
}
\value{
list containing input data, meta data, median calculated from the traits.
}
\description{
Median calculation
}
\examples{

library(plyr)
data("pe_data")
data("meta_data")

\dontrun{

# specify output directory
out_dir <- c()

test_median <- median_data_computation(data = pe_data, meta_data = meta_data,
output = out_dir)

}

}
